move calibration functions out to separate module
